requirements.txt
numpy>=1.24
opencv-python>=4.8
PyYAML>=6.0
scikit-learn>=1.4
matplotlib>=3.8
torch>=2.2
torchvision>=0.17
tqdm>=4.66
streamlit>=1.34
